requirements.txt

flask
flask-cors
google-auth
